backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.0
python-multipart==0.0.6
celery==5.5.3
redis==6.2.0
pytesseract==0.3.13
pillow==11.2.1
